requirements.txt
Django==5.1.3
djangorestframework==3.16.0
django-cors-headers==4.3.1

asgiref==3.8.1
daphne==4.0.0

requests==2.32.3
python-dotenv==1.0.1

psycopg2-binary==2.9.9
